+++ src/main/java/com/healthcare/model/records/DoctorRecord.java
@@ -3,7 +3,7 @@ package com.healthcare.model.records;
 import lombok.Builder;
 
 @Builder
-public record DoctorRecord(Long doctorId,
+public record DoctorRecord(Integer doctorId,
                            String name,
                            String speciality) {
 }
